requirements.txt
streamlit
sentence-transformers
PyPDF2
python-docx
faiss-cpu
rank-bm25
requests
python-dotenv